add validation scripts

+++ EKF/src/main.cpp
@@ -151,8 +151,8 @@ int main(int argc, char *argv[]) {
             out_file_ << measurement_pack_list[k].raw_measurements_(1) << "\t";
         } else if (measurement_pack_list[k].sensor_type_ == MeasurementPackage::RADAR) {
             // output the estimation in the cartesian coordinates
-            float ro = measurement_pack_list[k].raw_measurements_(0);
-            float phi = measurement_pack_list[k].raw_measurements_(1);
+            double ro = measurement_pack_list[k].raw_measurements_(0);
+            double phi = measurement_pack_list[k].raw_measurements_(1);
             out_file_ << ro * cos(phi) << "\t"; // p1_meas
             out_file_ << ro * sin(phi) << "\t"; // ps_meas
         }
